refactor the Indicators file structure and simplify the icons generation
(cherry picked from commit 2a5475a08dabd96c827ec33f9db49507e9448dcc)
