+++ src/pages/indicators/layout/main.css
